+++ src/main/java/ru/usque/pelican/repository/PelicanUserRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 public interface PelicanUserRepository extends CrudRepository<PelicanUser, Long>  {
     List<PelicanUser> findByEmail(String email);
-    PelicanUser findById(Long id);
+    PelicanUser findById(Integer id);
     List<PelicanUser> findByLogin(String login);
 }
